This is MyProject Documentation

Welcome to the documentation page.

Quick-Glimpse Tutorial

This tutorial just offered a quick glimpse on Julia's built-in documentation system, make sure to read the docs for more.

MyProject — Module

Simple dummy project

MyProject.ArithmeticSum — Method

ArithmeticSum(a)

does something